+++ src/renderer/src/store/thunks/index.ts
@@ -4,6 +4,7 @@ export * from '#/renderer/src/store/thunks/requests';
 export * from '#/renderer/src/store/thunks/modals';
 
 import type { AppDispatch } from '#/renderer/src/store/redux';
+import { setGeneralSettingsState } from '#/renderer/src/store/slices/settingsSlice';
 import { refreshCollections } from '#/renderer/src/store/thunks/collections';
 import { refreshEnvironments } from '#/renderer/src/store/thunks/environments';
 
@@ -13,4 +14,7 @@ import { refreshEnvironments } from '#/renderer/src/store/thunks/environments';
 export function initializeStore(dispatch: AppDispatch): void {
   void dispatch(refreshCollections());
   void dispatch(refreshEnvironments());
+  void window.api.getGeneralSettings().then((settings) => {
+    dispatch(setGeneralSettingsState(settings));
+  });
 }